+++ client/src/components/index/Slide.jsx
@@ -1,22 +1,30 @@
-import React from 'react';
-import 'swiper/swiper-bundle.css';
+import React from 'react'
 import { Swiper, SwiperSlide } from 'swiper/react';
-import SwiperCore, { Autoplay, Pagination } from 'swiper';
+import { Autoplay, Pagination } from 'swiper/modules';
+import 'swiper/css';
 
 const Slide = () => {
+  // SwiperCore.use([Autoplay])
   return (
     <Swiper
       spaceBetween={0}
       slidesPerView={1}
       loop={true}
-      autoplay={{ delay: 2000, disableOnInteraction: false }}
+      autoplay={{ delay: 3000, disableOnInteraction: false }}
       pagination={{ clickable: true }}
+      modules={[Autoplay, Pagination]}
     >
-      <SwiperSlide><img className='img' src={`${process.env.PUBLIC_URL}/img/slide01.png`} alt="" style={{ maxWidth: '100%', maxHeight: '600px' }} /></SwiperSlide>
-      <SwiperSlide><img className='img' src={`${process.env.PUBLIC_URL}/img/slide02.png`} alt="" style={{ maxWidth: '100%', maxHeight: '600px' }} /></SwiperSlide>
-      <SwiperSlide><img className='img' src={`${process.env.PUBLIC_URL}/img/slide03.png`} alt="" style={{ maxWidth: '100%', maxHeight: '600px' }} /></SwiperSlide>
+      <SwiperSlide>
+        <img className='img' src={`${process.env.PUBLIC_URL}/img/slide01.png`} alt="" style={{ maxWidth: '100%', maxHeight: '600px' }} />
+      </SwiperSlide>
+      <SwiperSlide>
+        <img className='img' src={`${process.env.PUBLIC_URL}/img/slide02.png`} alt="" style={{ maxWidth: '100%', maxHeight: '600px' }} />
+      </SwiperSlide>
+      <SwiperSlide>
+        <img className='img' src={`${process.env.PUBLIC_URL}/img/slide03.png`} alt="" style={{ maxWidth: '100%', maxHeight: '600px' }} />
+      </SwiperSlide>
     </Swiper>
-  );
-};
+  )
+}
 
-export default Slide;
+export default Slide
