+++ client/src/components/index/css/Home.css
@@ -1378,8 +1378,6 @@ button:not(:disabled),
   text-align: center !important;
 }
 
-/* rtl:begin:remove */
-
 .text-white {
   color: white !important;
   margin-bottom: 5%;
@@ -1611,13 +1609,10 @@ section {
   }
 }
 
-.color_white {color: white !important;}
-
-/* 스크롤 애니메이션 */
-.scroll_wrap {overflow: hidden;}
-.scroll_on {padding:50px 0;font-size: 24px;font-weight:700;text-align: center;opacity: 0;transition: all 1s;}
-.scroll_on.active {opacity: 1 !important;transform: translate(0, 0) !important;}
-.scroll_on.type_top {transform: translate(0, -50px);}
-.scroll_on.type_bottom {transform: translate(0, 50px);}
-.scroll_on.type_left {transform: translate(-50px, 0);}
-.scroll_on.type_right {transform: translate(50px, 0);}
+/* .index-body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 0;
+  flex-direction: column;
+} */
